GL changed (int)scale to (float)scale since value "0.99" from input was interpreted as int 0

+++ webclient/js/figures/drawable.js
@@ -30,7 +30,7 @@ class Drawable {
 
     scaleBy(scaleCoefficient) {
         // scale by all axes
-        const scale = parseInt(scaleCoefficient);
+        const scale = parseFloat(scaleCoefficient);
         this.scaleVec = [scale, scale, scale];
     }
